+++ src/main/java/database/model/Product.java
@@ -3,13 +3,13 @@ package database.model;
 public class Product {
     private String name;
     private Category category;
-    private Integer energy;
-    private Integer fat;
-    private Integer saturates;
-    private Integer carbohydrates;
-    private Integer sugars;
-    private Integer protein;
-    private Integer salt;
+    private Float energy;
+    private Float fat;
+    private Float saturates;
+    private Float carbohydrates;
+    private Float sugars;
+    private Float protein;
+    private Float salt;
 
     public String getName() {
         return name;
@@ -27,59 +27,59 @@ public class Product {
         this.category = category;
     }
 
-    public Integer getEnergy() {
+    public Float getEnergy() {
         return energy;
     }
 
-    public void setEnergy(Integer energy) {
+    public void setEnergy(Float energy) {
         this.energy = energy;
     }
 
-    public Integer getFat() {
+    public Float getFat() {
         return fat;
     }
 
-    public void setFat(Integer fat) {
+    public void setFat(Float fat) {
         this.fat = fat;
     }
 
-    public Integer getSaturates() {
+    public Float getSaturates() {
         return saturates;
     }
 
-    public void setSaturates(Integer saturates) {
+    public void setSaturates(Float saturates) {
         this.saturates = saturates;
     }
 
-    public Integer getCarbohydrates() {
+    public Float getCarbohydrates() {
         return carbohydrates;
     }
 
-    public void setCarbohydrates(Integer carbohydrates) {
+    public void setCarbohydrates(Float carbohydrates) {
         this.carbohydrates = carbohydrates;
     }
 
-    public Integer getSugars() {
+    public Float getSugars() {
         return sugars;
     }
 
-    public void setSugars(Integer sugars) {
+    public void setSugars(Float sugars) {
         this.sugars = sugars;
     }
 
-    public Integer getProtein() {
+    public Float getProtein() {
         return protein;
     }
 
-    public void setProtein(Integer protein) {
+    public void setProtein(Float protein) {
         this.protein = protein;
     }
 
-    public Integer getSalt() {
+    public Float getSalt() {
         return salt;
     }
 
-    public void setSalt(Integer salt) {
+    public void setSalt(Float salt) {
         this.salt = salt;
     }
 }
